Refactor add relationship form into a component
